fields graphs: a smaller pool per key -- int16 index planes, stage B's buffers inside it, freed on another key (same bytes)
- The face transforms' nearest-index planes are int16 (int32 only for a side of 2^15 or more; numpy's
  `face_distance` does the same): they were the largest statics a key's graphs hold (2 x 135 MB per
  224^3 block). The pair-check kernel and the torch pair checks widen them to int32 on load.
- Stage B's fixed-shape memory moves into graph A's pool (`_stage_b_ws`): both bands' union-find
  forest is computed IN graph A (two fewer eager launches per batch), and one Gaussian buffer is a
  static; both are allocated at the end of the capture, where they reuse the transients' dead pool
  memory (the pool did not grow). The Gaussian's other buffer is the dead forest in the eager path
  too (`gaussian3_box(out=)`), 90 MB less per block there.
- The eager cache is emptied before a capture (the pool no longer stacks on the first batch's
  cached blocks), and a batch of another key releases the graphs at once (`drop_other`), not only
  when that batch is graph-eligible.

Per key (224^3, RTX 5080), graph pool 0.771 -> 0.646 GB at B = 1, 2.301 -> 1.926 GB at B = 3
(0.81 / 2.43 GB before this round's kernels). Eager peak allocated per batch 0.712 -> 0.586 GB
(B = 1), 2.133 -> 1.756 GB (B = 3). Synthetic 1024^3 region, peak reserved: graphs 1.51 -> 1.31 GB
(B = 1), 4.45 -> 3.92 GB (B = 3); eager 0.99 -> 0.91 GB, 2.61 -> 2.36 GB. Per graphed batch 28 -> 25
API calls; device time unchanged (16.6 ms B = 1).

Bytes: fixtures bitwise equal to numpy on CUDA and CPU; the synthetic region byte-identical to HEAD
at batch 1 and 3, graphs on and off; new test: another key's batch releases the graphs.
tests/test_targets.py + tests/test_edt.py: 72 passed, 1 skipped.

+++ tests/test_edt.py
@@ -262,7 +262,7 @@ def test_the_face_pass_is_the_torch_face_distance():
                     r = r0.clone()
                     d, u, ix = E.face_edt(s, dy, dx, cap, reason=r, ev=ev, reach=2.5, code=5, torch_only=to)
                     out[to] = (d.view(torch.int32), u.view(torch.int32), ix, r)
-                    assert ix.dtype == torch.int32 and ix.shape == (3,) + tuple(batch.shape)
+                    assert ix.dtype == torch.int16 and ix.shape == (3,) + tuple(batch.shape)
                 for a, b in zip(out[False], out[True]):
                     assert torch.equal(a, b), (m.shape, cap)
                 assert (out[True][3] == 5).any()
